refactor: 입출력 관련 클래스를 view로 묶어주기
